refactor(og): update SEO metadata across multiple pages
