+++ pages/agenda/[id].vue
@@ -7,10 +7,7 @@ import {
   BreadcrumbPage,
   BreadcrumbSeparator,
 } from '@/components/ui/breadcrumb'
-import { ogImageUrl } from '~/constants'
 import type { ParsedAgendaData } from '~/types/agendas'
-
-const siteConfig = useSiteConfig()
 
 const route = useRoute()
 const id = route.params.id as string
@@ -48,7 +45,6 @@ useSeoMeta({
   author: () => speakerNames,
   ogTitle: `${agendaMarkdownData.value?.title} - ${speakerNames}`,
   ogDescription: agendaMarkdownData.value?.description,
-  ogImage: `${siteConfig.url}${ogImageUrl}`,
   twitterTitle: `${agendaMarkdownData.value?.title} - ${speakerNames}`,
   twitterDescription: agendaMarkdownData.value?.description,
 })
@@ -61,13 +57,15 @@ useSchemaOrg([
 
 const ogImageOptions = {
   component: 'OgImageTemplate',
-  title: agendaMarkdownData.value?.title,
-  description: agendaMarkdownData.value?.description,
-  author: speakerNames,
-  location: agendaMarkdownData.value?.location,
-  date: agendaMarkdownData.value?.date.split('T')[0],
-  startTime: agendaMarkdownData.value?.startTime,
-  endTime: agendaMarkdownData.value?.endTime,
+  props: {
+    title: agendaMarkdownData.value?.title,
+    description: agendaMarkdownData.value?.description,
+    author: speakerNames,
+    location: agendaMarkdownData.value?.location,
+    date: agendaMarkdownData.value?.date.split('T')[0],
+    startTime: agendaMarkdownData.value?.startTime,
+    endTime: agendaMarkdownData.value?.endTime,
+  },
 }
 
 defineOgImage(ogImageOptions)
